refactor: extract employee directory cells
