+++ apps/webapp/src/app/[locale]/(app)/settings/employees/translated-column-header.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { useTranslate } from "@tolgee/react";
+
+export function TranslatedColumnHeader({
+	className,
+	fallback,
+	nameKey,
+}: {
+	className?: string;
+	fallback: string;
+	nameKey: string;
+}) {
+	const { t } = useTranslate();
+
+	return <span className={className}>{t(nameKey, fallback)}</span>;
+}
